refactor(types): move types about api into separate file
